Hotfix: Update the text filter order
Now that the dataset URL can be edited when the WYSIWYG media embed plugin is used it embeds a media token with a URL. This creates an issue with the 'Convert URL to links' text format filter as this replaces all URLs (even those in tokens).

To work around this; the convert url filter should happen after the media filter has replaced the token.